making use of resources folder
