+++ package.json
@@ -19,5 +19,6 @@
   "devDependencies": {
     "@types/node": "^10.11.6",
     "typescript": "^3.1.2"
-  }
+  },
+  "files": ["./resources"]
 }
