+++ resources/myo-js.js
@@ -1,0 +1,17 @@
+/* Set the width of the side navigation to 250px and the left margin of the page content to 250px */
+function openNav() {
+    document.getElementById("mySidenav").style.width = "250px";
+    document.getElementById("mySidenav").style.overflow = "auto";
+    //document.getElementById("main").style.marginLeft = "250px";
+}
+
+/* Set the width of the side navigation to 0 and the left margin of the page content to 0 */
+function closeNav() {
+    document.getElementById("mySidenav").style.width = "0";
+    document.getElementById("mySidenav").style.overflow = "hidden";
+    //document.getElementById("main").style.marginLeft = "0";
+}
+
+document.addEventListener("DOMContentLoaded", () => {
+    closeNav();
+});
